3373cc930eb67559d29eeb65effe0d64dd3e0152 secp256k1-sys: update all symbols from 0.9.1 to 0.9.2 (Andrew Poelstra)
484e5d8d5b9d001af4ce5796b9837bbc704bebea fix incorrect FFI binding for pubkey_combine (Andrew Poelstra)

Pull request description:

  Fixes #669.

  Needs backport.

ACKs for top commit:
  Kixunil:
    ACK 3373cc930eb67559d29eeb65effe0d64dd3e0152

Tree-SHA512: 602a5baa8095cc744a341d64e300185bce26c8c56e7a538d3b17c7ca4c98cb3244217cd34169e1e69dc904a9f9f28ed75fe096ffa95ea42d1ad3456d395f7ce5

+++ src/key.rs
@@ -676,7 +676,7 @@ impl PublicKey {
                 ffi::secp256k1_context_no_precomp,
                 &mut ret,
                 ptrs.as_c_ptr(),
-                keys.len() as i32,
+                keys.len(),
             ) == 1
             {
                 Ok(PublicKey(ret))
